+++ docs/en/changes/changes.md
@@ -39,6 +39,7 @@
 * Fix Elasticsearch, MySQL, RabbitMQ dashboards typos and missing expressions.
 * BanyanDB: Zipkin Module set service as Entity for improving the query performance.
 * MQE: check the metrics value before do binary operation to improve robustness.
+* Replace workaround with Armeria native supported context path.
 
 #### UI